move integration tests
in order to use the cfg(test) code, this needs to be internal to the
crate
