+++ crates/hyperdrive-math/Cargo.toml
@@ -4,11 +4,6 @@ name = "hyperdrive-math"
 version = { workspace = true }
 edition = "2021"
 
-[[test]]
-
-name = "integration-tests"
-path = "tests/integration_tests.rs"
-
 [dependencies]
 
 # External dependencies
