+++ crates/hyperdrive-math/src/test_utils.rs
@@ -1,2 +1,3 @@
 #[cfg(test)]
 pub mod agent;
+mod integration_tests;
